hospital-mobile-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
supabase==1.1.1
pyjwt==2.8.0
cryptography==42.0.2
requests==2.31.0
python-multipart==0.0.6
email-validator==2.1.0
phonenumbers==8.13.27
pydantic==2.5.0
pydantic-settings==2.1.0